manage reply on trainee side
